refactor: move internal handler to a separate file
